@@ -6,7 +6,7 @@ part of 'base_view_model.dart';
 // StoreGenerator
 // **************************************************************************
 
-// ignore_for_file: non_constant_identifier_names, unnecessary_brace_in_string_interps, unnecessary_lambdas, prefer_expression_function_bodies, lines_longer_than_80_chars, avoid_as, avoid_annotating_with_dynamic
+// ignore_for_file: non_constant_identifier_names, unnecessary_brace_in_string_interps, unnecessary_lambdas, prefer_expression_function_bodies, lines_longer_than_80_chars, avoid_as, avoid_annotating_with_dynamic, no_leading_underscores_for_local_identifiers
 
 mixin _$BaseViewModel on _BaseViewModel, Store {
   Computed<bool>? _$isBusyComputed;
@@ -16,7 +16,8 @@ mixin _$BaseViewModel on _BaseViewModel, Store {
           Computed<bool>(() => super.isBusy, name: '_BaseViewModel.isBusy'))
       .value;
 
-  final _$_isBusyAtom = Atom(name: '_BaseViewModel._isBusy');
+  late final _$_isBusyAtom =
+      Atom(name: '_BaseViewModel._isBusy', context: context);
 
   @override
   bool get _isBusy {
@@ -31,8 +32,8 @@ mixin _$BaseViewModel on _BaseViewModel, Store {
     });
   }
 
-  final _$_BaseViewModelActionController =
-      ActionController(name: '_BaseViewModel');
+  late final _$_BaseViewModelActionController =
+      ActionController(name: '_BaseViewModel', context: context);
 
   @override
   void setBusy(bool state) {
