@@ -1,5 +1,7 @@
+// coverage:ignore-file
 // GENERATED CODE - DO NOT MODIFY BY HAND
-// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides
+// ignore_for_file: type=lint
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target, unnecessary_question_mark
 
 part of 'request_failure.dart';
 
@@ -10,27 +12,7 @@ part of 'request_failure.dart';
 T _$identity<T>(T value) => value;
 
 final _privateConstructorUsedError = UnsupportedError(
-    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more informations: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
-
-/// @nodoc
-class _$RequestFailureTearOff {
-  const _$RequestFailureTearOff();
-
-  _NetworkFailure networkError() {
-    return _NetworkFailure();
-  }
-
-  _Empty empty() {
-    return _Empty();
-  }
-
-  _ServerError serverError() {
-    return _ServerError();
-  }
-}
-
-/// @nodoc
-const $RequestFailure = _$RequestFailureTearOff();
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
 
 /// @nodoc
 mixin _$RequestFailure {
@@ -39,6 +21,13 @@ mixin _$RequestFailure {
     required TResult Function() networkError,
     required TResult Function() empty,
     required TResult Function() serverError,
+  }) =>
+      throw _privateConstructorUsedError;
+  @optionalTypeArgs
+  TResult? whenOrNull<TResult extends Object?>({
+    TResult? Function()? networkError,
+    TResult? Function()? empty,
+    TResult? Function()? serverError,
   }) =>
       throw _privateConstructorUsedError;
   @optionalTypeArgs
@@ -57,6 +46,13 @@ mixin _$RequestFailure {
   }) =>
       throw _privateConstructorUsedError;
   @optionalTypeArgs
+  TResult? mapOrNull<TResult extends Object?>({
+    TResult? Function(_NetworkFailure value)? networkError,
+    TResult? Function(_Empty value)? empty,
+    TResult? Function(_ServerError value)? serverError,
+  }) =>
+      throw _privateConstructorUsedError;
+  @optionalTypeArgs
   TResult maybeMap<TResult extends Object?>({
     TResult Function(_NetworkFailure value)? networkError,
     TResult Function(_Empty value)? empty,
@@ -70,36 +66,34 @@ mixin _$RequestFailure {
 abstract class $RequestFailureCopyWith<$Res> {
   factory $RequestFailureCopyWith(
           RequestFailure value, $Res Function(RequestFailure) then) =
-      _$RequestFailureCopyWithImpl<$Res>;
+      _$RequestFailureCopyWithImpl<$Res, RequestFailure>;
 }
 
 /// @nodoc
-class _$RequestFailureCopyWithImpl<$Res>
+class _$RequestFailureCopyWithImpl<$Res, $Val extends RequestFailure>
     implements $RequestFailureCopyWith<$Res> {
   _$RequestFailureCopyWithImpl(this._value, this._then);
 
-  final RequestFailure _value;
   // ignore: unused_field
-  final $Res Function(RequestFailure) _then;
+  final $Val _value;
+  // ignore: unused_field
+  final $Res Function($Val) _then;
 }
 
 /// @nodoc
-abstract class _$NetworkFailureCopyWith<$Res> {
-  factory _$NetworkFailureCopyWith(
-          _NetworkFailure value, $Res Function(_NetworkFailure) then) =
-      __$NetworkFailureCopyWithImpl<$Res>;
+abstract class _$$_NetworkFailureCopyWith<$Res> {
+  factory _$$_NetworkFailureCopyWith(
+          _$_NetworkFailure value, $Res Function(_$_NetworkFailure) then) =
+      __$$_NetworkFailureCopyWithImpl<$Res>;
 }
 
 /// @nodoc
-class __$NetworkFailureCopyWithImpl<$Res>
-    extends _$RequestFailureCopyWithImpl<$Res>
-    implements _$NetworkFailureCopyWith<$Res> {
-  __$NetworkFailureCopyWithImpl(
-      _NetworkFailure _value, $Res Function(_NetworkFailure) _then)
-      : super(_value, (v) => _then(v as _NetworkFailure));
-
-  @override
-  _NetworkFailure get _value => super._value as _NetworkFailure;
+class __$$_NetworkFailureCopyWithImpl<$Res>
+    extends _$RequestFailureCopyWithImpl<$Res, _$_NetworkFailure>
+    implements _$$_NetworkFailureCopyWith<$Res> {
+  __$$_NetworkFailureCopyWithImpl(
+      _$_NetworkFailure _value, $Res Function(_$_NetworkFailure) _then)
+      : super(_value, _then);
 }
 
 /// @nodoc
@@ -114,7 +108,8 @@ class _$_NetworkFailure implements _NetworkFailure {
 
   @override
   bool operator ==(dynamic other) {
-    return identical(this, other) || (other is _NetworkFailure);
+    return identical(this, other) ||
+        (other.runtimeType == runtimeType && other is _$_NetworkFailure);
   }
 
   @override
@@ -128,6 +123,16 @@ class _$_NetworkFailure implements _NetworkFailure {
     required TResult Function() serverError,
   }) {
     return networkError();
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult? whenOrNull<TResult extends Object?>({
+    TResult? Function()? networkError,
+    TResult? Function()? empty,
+    TResult? Function()? serverError,
+  }) {
+    return networkError?.call();
   }
 
   @override
@@ -156,6 +161,16 @@ class _$_NetworkFailure implements _NetworkFailure {
 
   @override
   @optionalTypeArgs
+  TResult? mapOrNull<TResult extends Object?>({
+    TResult? Function(_NetworkFailure value)? networkError,
+    TResult? Function(_Empty value)? empty,
+    TResult? Function(_ServerError value)? serverError,
+  }) {
+    return networkError?.call(this);
+  }
+
+  @override
+  @optionalTypeArgs
   TResult maybeMap<TResult extends Object?>({
     TResult Function(_NetworkFailure value)? networkError,
     TResult Function(_Empty value)? empty,
@@ -174,19 +189,17 @@ abstract class _NetworkFailure implements RequestFailure {
 }
 
 /// @nodoc
-abstract class _$EmptyCopyWith<$Res> {
-  factory _$EmptyCopyWith(_Empty value, $Res Function(_Empty) then) =
-      __$EmptyCopyWithImpl<$Res>;
+abstract class _$$_EmptyCopyWith<$Res> {
+  factory _$$_EmptyCopyWith(_$_Empty value, $Res Function(_$_Empty) then) =
+      __$$_EmptyCopyWithImpl<$Res>;
 }
 
 /// @nodoc
-class __$EmptyCopyWithImpl<$Res> extends _$RequestFailureCopyWithImpl<$Res>
-    implements _$EmptyCopyWith<$Res> {
-  __$EmptyCopyWithImpl(_Empty _value, $Res Function(_Empty) _then)
-      : super(_value, (v) => _then(v as _Empty));
-
-  @override
-  _Empty get _value => super._value as _Empty;
+class __$$_EmptyCopyWithImpl<$Res>
+    extends _$RequestFailureCopyWithImpl<$Res, _$_Empty>
+    implements _$$_EmptyCopyWith<$Res> {
+  __$$_EmptyCopyWithImpl(_$_Empty _value, $Res Function(_$_Empty) _then)
+      : super(_value, _then);
 }
 
 /// @nodoc
@@ -201,7 +214,8 @@ class _$_Empty implements _Empty {
 
   @override
   bool operator ==(dynamic other) {
-    return identical(this, other) || (other is _Empty);
+    return identical(this, other) ||
+        (other.runtimeType == runtimeType && other is _$_Empty);
   }
 
   @override
@@ -215,6 +229,16 @@ class _$_Empty implements _Empty {
     required TResult Function() serverError,
   }) {
     return empty();
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult? whenOrNull<TResult extends Object?>({
+    TResult? Function()? networkError,
+    TResult? Function()? empty,
+    TResult? Function()? serverError,
+  }) {
+    return empty?.call();
   }
 
   @override
@@ -243,6 +267,16 @@ class _$_Empty implements _Empty {
 
   @override
   @optionalTypeArgs
+  TResult? mapOrNull<TResult extends Object?>({
+    TResult? Function(_NetworkFailure value)? networkError,
+    TResult? Function(_Empty value)? empty,
+    TResult? Function(_ServerError value)? serverError,
+  }) {
+    return empty?.call(this);
+  }
+
+  @override
+  @optionalTypeArgs
   TResult maybeMap<TResult extends Object?>({
     TResult Function(_NetworkFailure value)? networkError,
     TResult Function(_Empty value)? empty,
@@ -261,22 +295,19 @@ abstract class _Empty implements RequestFailure {
 }
 
 /// @nodoc
-abstract class _$ServerErrorCopyWith<$Res> {
-  factory _$ServerErrorCopyWith(
-          _ServerError value, $Res Function(_ServerError) then) =
-      __$ServerErrorCopyWithImpl<$Res>;
+abstract class _$$_ServerErrorCopyWith<$Res> {
+  factory _$$_ServerErrorCopyWith(
+          _$_ServerError value, $Res Function(_$_ServerError) then) =
+      __$$_ServerErrorCopyWithImpl<$Res>;
 }
 
 /// @nodoc
-class __$ServerErrorCopyWithImpl<$Res>
-    extends _$RequestFailureCopyWithImpl<$Res>
-    implements _$ServerErrorCopyWith<$Res> {
-  __$ServerErrorCopyWithImpl(
-      _ServerError _value, $Res Function(_ServerError) _then)
-      : super(_value, (v) => _then(v as _ServerError));
-
-  @override
-  _ServerError get _value => super._value as _ServerError;
+class __$$_ServerErrorCopyWithImpl<$Res>
+    extends _$RequestFailureCopyWithImpl<$Res, _$_ServerError>
+    implements _$$_ServerErrorCopyWith<$Res> {
+  __$$_ServerErrorCopyWithImpl(
+      _$_ServerError _value, $Res Function(_$_ServerError) _then)
+      : super(_value, _then);
 }
 
 /// @nodoc
@@ -291,7 +322,8 @@ class _$_ServerError implements _ServerError {
 
   @override
   bool operator ==(dynamic other) {
-    return identical(this, other) || (other is _ServerError);
+    return identical(this, other) ||
+        (other.runtimeType == runtimeType && other is _$_ServerError);
   }
 
   @override
@@ -305,6 +337,16 @@ class _$_ServerError implements _ServerError {
     required TResult Function() serverError,
   }) {
     return serverError();
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult? whenOrNull<TResult extends Object?>({
+    TResult? Function()? networkError,
+    TResult? Function()? empty,
+    TResult? Function()? serverError,
+  }) {
+    return serverError?.call();
   }
 
   @override
@@ -329,6 +371,16 @@ class _$_ServerError implements _ServerError {
     required TResult Function(_ServerError value) serverError,
   }) {
     return serverError(this);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult? mapOrNull<TResult extends Object?>({
+    TResult? Function(_NetworkFailure value)? networkError,
+    TResult? Function(_Empty value)? empty,
+    TResult? Function(_ServerError value)? serverError,
+  }) {
+    return serverError?.call(this);
   }
 
   @override
