@@ -2,13 +2,13 @@ import 'package:flutter_package/src/domain/packages/entities/pubspec.dart';
 
 class Version {
   final String version;
-  final String archive_url;
+  final String archiveUrl;
   final String published;
   final Pubspec pubspec;
 
   Version(
       {required this.version,
-      required this.archive_url,
+      required this.archiveUrl,
       required this.pubspec,
       required this.published});
 
@@ -16,7 +16,7 @@ class Version {
     return {
       'version': this.version,
       'published': this.published,
-      'archive_url': this.archive_url,
+      'archive_url': this.archiveUrl,
       'pubspec': this.pubspec,
     };
   }
