@@ -36,7 +36,7 @@ class BaseSearchRepository extends ISearchRepository {
         return left(RequestFailure.networkError());
       }
 
-      if (e.type == DioErrorType.connectTimeout ||
+      if (e.type == DioErrorType.connectionTimeout ||
           e.type == DioErrorType.receiveTimeout ||
           e.type == DioErrorType.sendTimeout) {
         return left(RequestFailure.serverError());
