@@ -154,7 +154,7 @@ class PackagesScreenState extends State<PackagesScreen>
     return ConfigBuilder(builder: (_, theme) {
       return Scaffold(
         appBar: AppBar(
-          brightness: theme.brightness,
+          systemOverlayStyle: theme.appBarTheme.systemOverlayStyle,
           centerTitle: false,
           actions: <Widget>[
             IconButton(
