@@ -1,7 +1,8 @@
-// Mocks generated by Mockito 5.0.15 from annotations
+// Mocks generated by Mockito 5.3.2 from annotations
 // in flutter_package/test/src/infrastructure/mocks.dart.
 // Do not manually edit this file.
 
+// ignore_for_file: no_leading_underscores_for_library_prefixes
 import 'dart:async' as _i5;
 
 import 'package:dartz/dartz.dart' as _i3;
@@ -19,6 +20,7 @@ import 'package:flutter_package/src/domain/search/i_search_repository.dart'
     as _i9;
 import 'package:mockito/mockito.dart' as _i1;
 
+// ignore_for_file: type=lint
 // ignore_for_file: avoid_redundant_argument_values
 // ignore_for_file: avoid_setters_without_getters
 // ignore_for_file: comment_references
@@ -26,21 +28,79 @@ import 'package:mockito/mockito.dart' as _i1;
 // ignore_for_file: invalid_use_of_visible_for_testing_member
 // ignore_for_file: prefer_const_constructors
 // ignore_for_file: unnecessary_parenthesis
+// ignore_for_file: camel_case_types
+// ignore_for_file: subtype_of_sealed_class
 
-class _FakeDio_0 extends _i1.Fake implements _i2.Dio {}
+class _FakeDio_0 extends _i1.SmartFake implements _i2.Dio {
+  _FakeDio_0(
+    Object parent,
+    Invocation parentInvocation,
+  ) : super(
+          parent,
+          parentInvocation,
+        );
+}
 
-class _FakeEither_1<L, R> extends _i1.Fake implements _i3.Either<L, R> {}
+class _FakeEither_1<L, R> extends _i1.SmartFake implements _i3.Either<L, R> {
+  _FakeEither_1(
+    Object parent,
+    Invocation parentInvocation,
+  ) : super(
+          parent,
+          parentInvocation,
+        );
+}
 
-class _FakeBaseOptions_2 extends _i1.Fake implements _i2.BaseOptions {}
+class _FakeBaseOptions_2 extends _i1.SmartFake implements _i2.BaseOptions {
+  _FakeBaseOptions_2(
+    Object parent,
+    Invocation parentInvocation,
+  ) : super(
+          parent,
+          parentInvocation,
+        );
+}
 
-class _FakeHttpClientAdapter_3 extends _i1.Fake
-    implements _i2.HttpClientAdapter {}
+class _FakeHttpClientAdapter_3 extends _i1.SmartFake
+    implements _i2.HttpClientAdapter {
+  _FakeHttpClientAdapter_3(
+    Object parent,
+    Invocation parentInvocation,
+  ) : super(
+          parent,
+          parentInvocation,
+        );
+}
 
-class _FakeTransformer_4 extends _i1.Fake implements _i2.Transformer {}
+class _FakeTransformer_4 extends _i1.SmartFake implements _i2.Transformer {
+  _FakeTransformer_4(
+    Object parent,
+    Invocation parentInvocation,
+  ) : super(
+          parent,
+          parentInvocation,
+        );
+}
 
-class _FakeInterceptors_5 extends _i1.Fake implements _i2.Interceptors {}
+class _FakeInterceptors_5 extends _i1.SmartFake implements _i2.Interceptors {
+  _FakeInterceptors_5(
+    Object parent,
+    Invocation parentInvocation,
+  ) : super(
+          parent,
+          parentInvocation,
+        );
+}
 
-class _FakeResponse_6<T> extends _i1.Fake implements _i2.Response<T> {}
+class _FakeResponse_6<T> extends _i1.SmartFake implements _i2.Response<T> {
+  _FakeResponse_6(
+    Object parent,
+    Invocation parentInvocation,
+  ) : super(
+          parent,
+          parentInvocation,
+        );
+}
 
 /// A class which mocks [IPackageRepository].
 ///
@@ -52,47 +112,97 @@ class MockIPackageRepository extends _i1.Mock
   }
 
   @override
-  String get path =>
-      (super.noSuchMethod(Invocation.getter(#path), returnValue: '') as String);
+  String get path => (super.noSuchMethod(
+        Invocation.getter(#path),
+        returnValue: '',
+      ) as String);
   @override
-  _i2.Dio get dio =>
-      (super.noSuchMethod(Invocation.getter(#dio), returnValue: _FakeDio_0())
-          as _i2.Dio);
+  _i2.Dio get dio => (super.noSuchMethod(
+        Invocation.getter(#dio),
+        returnValue: _FakeDio_0(
+          this,
+          Invocation.getter(#dio),
+        ),
+      ) as _i2.Dio);
   @override
   _i5.Future<_i3.Either<_i6.RequestFailure, List<_i7.Package>>> getPackages(
-          {int? page}) =>
-      (super.noSuchMethod(Invocation.method(#getPackages, [], {#page: page}),
-              returnValue: Future<
-                      _i3.Either<_i6.RequestFailure, List<_i7.Package>>>.value(
-                  _FakeEither_1<_i6.RequestFailure, List<_i7.Package>>()))
-          as _i5.Future<_i3.Either<_i6.RequestFailure, List<_i7.Package>>>);
+          {required int? page}) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #getPackages,
+          [],
+          {#page: page},
+        ),
+        returnValue:
+            _i5.Future<_i3.Either<_i6.RequestFailure, List<_i7.Package>>>.value(
+                _FakeEither_1<_i6.RequestFailure, List<_i7.Package>>(
+          this,
+          Invocation.method(
+            #getPackages,
+            [],
+            {#page: page},
+          ),
+        )),
+      ) as _i5.Future<_i3.Either<_i6.RequestFailure, List<_i7.Package>>>);
   @override
   _i5.Future<_i3.Either<_i6.RequestFailure, String>> getPublisher(
-          {String? packageName}) =>
+          {required String? packageName}) =>
       (super.noSuchMethod(
-              Invocation.method(#getPublisher, [], {#packageName: packageName}),
-              returnValue: Future<_i3.Either<_i6.RequestFailure, String>>.value(
-                  _FakeEither_1<_i6.RequestFailure, String>()))
-          as _i5.Future<_i3.Either<_i6.RequestFailure, String>>);
+        Invocation.method(
+          #getPublisher,
+          [],
+          {#packageName: packageName},
+        ),
+        returnValue: _i5.Future<_i3.Either<_i6.RequestFailure, String>>.value(
+            _FakeEither_1<_i6.RequestFailure, String>(
+          this,
+          Invocation.method(
+            #getPublisher,
+            [],
+            {#packageName: packageName},
+          ),
+        )),
+      ) as _i5.Future<_i3.Either<_i6.RequestFailure, String>>);
   @override
   _i5.Future<_i3.Either<_i6.RequestFailure, _i7.Package>> getPackageName(
-          {String? packageName}) =>
+          {required String? packageName}) =>
       (super.noSuchMethod(
-          Invocation.method(#getPackageName, [], {#packageName: packageName}),
-          returnValue:
-              Future<_i3.Either<_i6.RequestFailure, _i7.Package>>.value(
-                  _FakeEither_1<_i6.RequestFailure, _i7.Package>())) as _i5
-          .Future<_i3.Either<_i6.RequestFailure, _i7.Package>>);
+        Invocation.method(
+          #getPackageName,
+          [],
+          {#packageName: packageName},
+        ),
+        returnValue:
+            _i5.Future<_i3.Either<_i6.RequestFailure, _i7.Package>>.value(
+                _FakeEither_1<_i6.RequestFailure, _i7.Package>(
+          this,
+          Invocation.method(
+            #getPackageName,
+            [],
+            {#packageName: packageName},
+          ),
+        )),
+      ) as _i5.Future<_i3.Either<_i6.RequestFailure, _i7.Package>>);
   @override
   _i5.Future<_i3.Either<_i6.RequestFailure, _i8.Metric>> getMetricPackage(
-          {String? package}) =>
+          {required String? package}) =>
       (super.noSuchMethod(
-          Invocation.method(#getMetricPackage, [], {#package: package}),
-          returnValue: Future<_i3.Either<_i6.RequestFailure, _i8.Metric>>.value(
-              _FakeEither_1<_i6.RequestFailure, _i8.Metric>())) as _i5
-          .Future<_i3.Either<_i6.RequestFailure, _i8.Metric>>);
-  @override
-  String toString() => super.toString();
+        Invocation.method(
+          #getMetricPackage,
+          [],
+          {#package: package},
+        ),
+        returnValue:
+            _i5.Future<_i3.Either<_i6.RequestFailure, _i8.Metric>>.value(
+                _FakeEither_1<_i6.RequestFailure, _i8.Metric>(
+          this,
+          Invocation.method(
+            #getMetricPackage,
+            [],
+            {#package: package},
+          ),
+        )),
+      ) as _i5.Future<_i3.Either<_i6.RequestFailure, _i8.Metric>>);
 }
 
 /// A class which mocks [ISearchRepository].
@@ -104,23 +214,46 @@ class MockISearchRepository extends _i1.Mock implements _i9.ISearchRepository {
   }
 
   @override
-  String get path =>
-      (super.noSuchMethod(Invocation.getter(#path), returnValue: '') as String);
+  String get path => (super.noSuchMethod(
+        Invocation.getter(#path),
+        returnValue: '',
+      ) as String);
   @override
-  _i2.Dio get dio =>
-      (super.noSuchMethod(Invocation.getter(#dio), returnValue: _FakeDio_0())
-          as _i2.Dio);
+  _i2.Dio get dio => (super.noSuchMethod(
+        Invocation.getter(#dio),
+        returnValue: _FakeDio_0(
+          this,
+          Invocation.getter(#dio),
+        ),
+      ) as _i2.Dio);
   @override
-  _i5.Future<_i3.Either<_i6.RequestFailure, List<String>>> searchPackage(
-          {int? page, String? query}) =>
+  _i5.Future<_i3.Either<_i6.RequestFailure, List<String>>> searchPackage({
+    required int? page,
+    required String? query,
+  }) =>
       (super.noSuchMethod(
-          Invocation.method(#searchPackage, [], {#page: page, #query: query}),
-          returnValue:
-              Future<_i3.Either<_i6.RequestFailure, List<String>>>.value(
-                  _FakeEither_1<_i6.RequestFailure, List<String>>())) as _i5
-          .Future<_i3.Either<_i6.RequestFailure, List<String>>>);
-  @override
-  String toString() => super.toString();
+        Invocation.method(
+          #searchPackage,
+          [],
+          {
+            #page: page,
+            #query: query,
+          },
+        ),
+        returnValue:
+            _i5.Future<_i3.Either<_i6.RequestFailure, List<String>>>.value(
+                _FakeEither_1<_i6.RequestFailure, List<String>>(
+          this,
+          Invocation.method(
+            #searchPackage,
+            [],
+            {
+              #page: page,
+              #query: query,
+            },
+          ),
+        )),
+      ) as _i5.Future<_i3.Either<_i6.RequestFailure, List<String>>>);
 }
 
 /// A class which mocks [IAdvancedService].
@@ -132,16 +265,32 @@ class MockIAdvancedService extends _i1.Mock implements _i10.IAdvancedService {
   }
 
   @override
-  _i5.Future<_i3.Either<_i6.RequestFailure, String>> readFile(
-          {String? gitPath, String? fileName}) =>
+  _i5.Future<_i3.Either<_i6.RequestFailure, String>> readFile({
+    required String? gitPath,
+    required String? fileName,
+  }) =>
       (super.noSuchMethod(
-              Invocation.method(
-                  #readFile, [], {#gitPath: gitPath, #fileName: fileName}),
-              returnValue: Future<_i3.Either<_i6.RequestFailure, String>>.value(
-                  _FakeEither_1<_i6.RequestFailure, String>()))
-          as _i5.Future<_i3.Either<_i6.RequestFailure, String>>);
-  @override
-  String toString() => super.toString();
+        Invocation.method(
+          #readFile,
+          [],
+          {
+            #gitPath: gitPath,
+            #fileName: fileName,
+          },
+        ),
+        returnValue: _i5.Future<_i3.Either<_i6.RequestFailure, String>>.value(
+            _FakeEither_1<_i6.RequestFailure, String>(
+          this,
+          Invocation.method(
+            #readFile,
+            [],
+            {
+              #gitPath: gitPath,
+              #fileName: fileName,
+            },
+          ),
+        )),
+      ) as _i5.Future<_i3.Either<_i6.RequestFailure, String>>);
 }
 
 /// A class which mocks [Dio].
@@ -153,341 +302,676 @@ class MockApi extends _i1.Mock implements _i2.Dio {
   }
 
   @override
-  _i2.BaseOptions get options =>
-      (super.noSuchMethod(Invocation.getter(#options),
-          returnValue: _FakeBaseOptions_2()) as _i2.BaseOptions);
+  _i2.BaseOptions get options => (super.noSuchMethod(
+        Invocation.getter(#options),
+        returnValue: _FakeBaseOptions_2(
+          this,
+          Invocation.getter(#options),
+        ),
+      ) as _i2.BaseOptions);
   @override
-  set options(_i2.BaseOptions? _options) =>
-      super.noSuchMethod(Invocation.setter(#options, _options),
-          returnValueForMissingStub: null);
+  set options(_i2.BaseOptions? _options) => super.noSuchMethod(
+        Invocation.setter(
+          #options,
+          _options,
+        ),
+        returnValueForMissingStub: null,
+      );
   @override
-  _i2.HttpClientAdapter get httpClientAdapter =>
-      (super.noSuchMethod(Invocation.getter(#httpClientAdapter),
-          returnValue: _FakeHttpClientAdapter_3()) as _i2.HttpClientAdapter);
+  _i2.HttpClientAdapter get httpClientAdapter => (super.noSuchMethod(
+        Invocation.getter(#httpClientAdapter),
+        returnValue: _FakeHttpClientAdapter_3(
+          this,
+          Invocation.getter(#httpClientAdapter),
+        ),
+      ) as _i2.HttpClientAdapter);
   @override
-  set httpClientAdapter(_i2.HttpClientAdapter? _httpClientAdapter) => super
-      .noSuchMethod(Invocation.setter(#httpClientAdapter, _httpClientAdapter),
-          returnValueForMissingStub: null);
+  set httpClientAdapter(_i2.HttpClientAdapter? _httpClientAdapter) =>
+      super.noSuchMethod(
+        Invocation.setter(
+          #httpClientAdapter,
+          _httpClientAdapter,
+        ),
+        returnValueForMissingStub: null,
+      );
   @override
-  _i2.Transformer get transformer =>
-      (super.noSuchMethod(Invocation.getter(#transformer),
-          returnValue: _FakeTransformer_4()) as _i2.Transformer);
+  _i2.Transformer get transformer => (super.noSuchMethod(
+        Invocation.getter(#transformer),
+        returnValue: _FakeTransformer_4(
+          this,
+          Invocation.getter(#transformer),
+        ),
+      ) as _i2.Transformer);
   @override
-  set transformer(_i2.Transformer? _transformer) =>
-      super.noSuchMethod(Invocation.setter(#transformer, _transformer),
-          returnValueForMissingStub: null);
+  set transformer(_i2.Transformer? _transformer) => super.noSuchMethod(
+        Invocation.setter(
+          #transformer,
+          _transformer,
+        ),
+        returnValueForMissingStub: null,
+      );
   @override
-  _i2.Interceptors get interceptors =>
-      (super.noSuchMethod(Invocation.getter(#interceptors),
-          returnValue: _FakeInterceptors_5()) as _i2.Interceptors);
+  _i2.Interceptors get interceptors => (super.noSuchMethod(
+        Invocation.getter(#interceptors),
+        returnValue: _FakeInterceptors_5(
+          this,
+          Invocation.getter(#interceptors),
+        ),
+      ) as _i2.Interceptors);
   @override
-  void close({bool? force = false}) =>
-      super.noSuchMethod(Invocation.method(#close, [], {#force: force}),
-          returnValueForMissingStub: null);
+  void close({bool? force = false}) => super.noSuchMethod(
+        Invocation.method(
+          #close,
+          [],
+          {#force: force},
+        ),
+        returnValueForMissingStub: null,
+      );
   @override
-  _i5.Future<_i2.Response<T>> get<T>(String? path,
-          {Map<String, dynamic>? queryParameters,
-          _i2.Options? options,
-          _i2.CancelToken? cancelToken,
-          _i2.ProgressCallback? onReceiveProgress}) =>
+  _i5.Future<_i2.Response<T>> get<T>(
+    String? path, {
+    Object? data,
+    Map<String, dynamic>? queryParameters,
+    _i2.Options? options,
+    _i2.CancelToken? cancelToken,
+    _i2.ProgressCallback? onReceiveProgress,
+  }) =>
       (super.noSuchMethod(
-              Invocation.method(#get, [
-                path
-              ], {
-                #queryParameters: queryParameters,
-                #options: options,
-                #cancelToken: cancelToken,
-                #onReceiveProgress: onReceiveProgress
-              }),
-              returnValue: Future<_i2.Response<T>>.value(_FakeResponse_6<T>()))
-          as _i5.Future<_i2.Response<T>>);
+        Invocation.method(
+          #get,
+          [path],
+          {
+            #data: data,
+            #queryParameters: queryParameters,
+            #options: options,
+            #cancelToken: cancelToken,
+            #onReceiveProgress: onReceiveProgress,
+          },
+        ),
+        returnValue: _i5.Future<_i2.Response<T>>.value(_FakeResponse_6<T>(
+          this,
+          Invocation.method(
+            #get,
+            [path],
+            {
+              #data: data,
+              #queryParameters: queryParameters,
+              #options: options,
+              #cancelToken: cancelToken,
+              #onReceiveProgress: onReceiveProgress,
+            },
+          ),
+        )),
+      ) as _i5.Future<_i2.Response<T>>);
   @override
-  _i5.Future<_i2.Response<T>> getUri<T>(Uri? uri,
-          {_i2.Options? options,
-          _i2.CancelToken? cancelToken,
-          _i2.ProgressCallback? onReceiveProgress}) =>
+  _i5.Future<_i2.Response<T>> getUri<T>(
+    Uri? uri, {
+    Object? data,
+    _i2.Options? options,
+    _i2.CancelToken? cancelToken,
+    _i2.ProgressCallback? onReceiveProgress,
+  }) =>
       (super.noSuchMethod(
-              Invocation.method(#getUri, [
-                uri
-              ], {
-                #options: options,
-                #cancelToken: cancelToken,
-                #onReceiveProgress: onReceiveProgress
-              }),
-              returnValue: Future<_i2.Response<T>>.value(_FakeResponse_6<T>()))
-          as _i5.Future<_i2.Response<T>>);
+        Invocation.method(
+          #getUri,
+          [uri],
+          {
+            #data: data,
+            #options: options,
+            #cancelToken: cancelToken,
+            #onReceiveProgress: onReceiveProgress,
+          },
+        ),
+        returnValue: _i5.Future<_i2.Response<T>>.value(_FakeResponse_6<T>(
+          this,
+          Invocation.method(
+            #getUri,
+            [uri],
+            {
+              #data: data,
+              #options: options,
+              #cancelToken: cancelToken,
+              #onReceiveProgress: onReceiveProgress,
+            },
+          ),
+        )),
+      ) as _i5.Future<_i2.Response<T>>);
   @override
-  _i5.Future<_i2.Response<T>> post<T>(String? path,
-          {dynamic data,
-          Map<String, dynamic>? queryParameters,
-          _i2.Options? options,
-          _i2.CancelToken? cancelToken,
-          _i2.ProgressCallback? onSendProgress,
-          _i2.ProgressCallback? onReceiveProgress}) =>
+  _i5.Future<_i2.Response<T>> post<T>(
+    String? path, {
+    Object? data,
+    Map<String, dynamic>? queryParameters,
+    _i2.Options? options,
+    _i2.CancelToken? cancelToken,
+    _i2.ProgressCallback? onSendProgress,
+    _i2.ProgressCallback? onReceiveProgress,
+  }) =>
       (super.noSuchMethod(
-              Invocation.method(#post, [
-                path
-              ], {
-                #data: data,
-                #queryParameters: queryParameters,
-                #options: options,
-                #cancelToken: cancelToken,
-                #onSendProgress: onSendProgress,
-                #onReceiveProgress: onReceiveProgress
-              }),
-              returnValue: Future<_i2.Response<T>>.value(_FakeResponse_6<T>()))
-          as _i5.Future<_i2.Response<T>>);
+        Invocation.method(
+          #post,
+          [path],
+          {
+            #data: data,
+            #queryParameters: queryParameters,
+            #options: options,
+            #cancelToken: cancelToken,
+            #onSendProgress: onSendProgress,
+            #onReceiveProgress: onReceiveProgress,
+          },
+        ),
+        returnValue: _i5.Future<_i2.Response<T>>.value(_FakeResponse_6<T>(
+          this,
+          Invocation.method(
+            #post,
+            [path],
+            {
+              #data: data,
+              #queryParameters: queryParameters,
+              #options: options,
+              #cancelToken: cancelToken,
+              #onSendProgress: onSendProgress,
+              #onReceiveProgress: onReceiveProgress,
+            },
+          ),
+        )),
+      ) as _i5.Future<_i2.Response<T>>);
   @override
-  _i5.Future<_i2.Response<T>> postUri<T>(Uri? uri,
-          {dynamic data,
-          _i2.Options? options,
-          _i2.CancelToken? cancelToken,
-          _i2.ProgressCallback? onSendProgress,
-          _i2.ProgressCallback? onReceiveProgress}) =>
+  _i5.Future<_i2.Response<T>> postUri<T>(
+    Uri? uri, {
+    Object? data,
+    _i2.Options? options,
+    _i2.CancelToken? cancelToken,
+    _i2.ProgressCallback? onSendProgress,
+    _i2.ProgressCallback? onReceiveProgress,
+  }) =>
       (super.noSuchMethod(
-              Invocation.method(#postUri, [
-                uri
-              ], {
-                #data: data,
-                #options: options,
-                #cancelToken: cancelToken,
-                #onSendProgress: onSendProgress,
-                #onReceiveProgress: onReceiveProgress
-              }),
-              returnValue: Future<_i2.Response<T>>.value(_FakeResponse_6<T>()))
-          as _i5.Future<_i2.Response<T>>);
+        Invocation.method(
+          #postUri,
+          [uri],
+          {
+            #data: data,
+            #options: options,
+            #cancelToken: cancelToken,
+            #onSendProgress: onSendProgress,
+            #onReceiveProgress: onReceiveProgress,
+          },
+        ),
+        returnValue: _i5.Future<_i2.Response<T>>.value(_FakeResponse_6<T>(
+          this,
+          Invocation.method(
+            #postUri,
+            [uri],
+            {
+              #data: data,
+              #options: options,
+              #cancelToken: cancelToken,
+              #onSendProgress: onSendProgress,
+              #onReceiveProgress: onReceiveProgress,
+            },
+          ),
+        )),
+      ) as _i5.Future<_i2.Response<T>>);
   @override
-  _i5.Future<_i2.Response<T>> put<T>(String? path,
-          {dynamic data,
-          Map<String, dynamic>? queryParameters,
-          _i2.Options? options,
-          _i2.CancelToken? cancelToken,
-          _i2.ProgressCallback? onSendProgress,
-          _i2.ProgressCallback? onReceiveProgress}) =>
+  _i5.Future<_i2.Response<T>> put<T>(
+    String? path, {
+    Object? data,
+    Map<String, dynamic>? queryParameters,
+    _i2.Options? options,
+    _i2.CancelToken? cancelToken,
+    _i2.ProgressCallback? onSendProgress,
+    _i2.ProgressCallback? onReceiveProgress,
+  }) =>
       (super.noSuchMethod(
-              Invocation.method(#put, [
-                path
-              ], {
-                #data: data,
-                #queryParameters: queryParameters,
-                #options: options,
-                #cancelToken: cancelToken,
-                #onSendProgress: onSendProgress,
-                #onReceiveProgress: onReceiveProgress
-              }),
-              returnValue: Future<_i2.Response<T>>.value(_FakeResponse_6<T>()))
-          as _i5.Future<_i2.Response<T>>);
+        Invocation.method(
+          #put,
+          [path],
+          {
+            #data: data,
+            #queryParameters: queryParameters,
+            #options: options,
+            #cancelToken: cancelToken,
+            #onSendProgress: onSendProgress,
+            #onReceiveProgress: onReceiveProgress,
+          },
+        ),
+        returnValue: _i5.Future<_i2.Response<T>>.value(_FakeResponse_6<T>(
+          this,
+          Invocation.method(
+            #put,
+            [path],
+            {
+              #data: data,
+              #queryParameters: queryParameters,
+              #options: options,
+              #cancelToken: cancelToken,
+              #onSendProgress: onSendProgress,
+              #onReceiveProgress: onReceiveProgress,
+            },
+          ),
+        )),
+      ) as _i5.Future<_i2.Response<T>>);
   @override
-  _i5.Future<_i2.Response<T>> putUri<T>(Uri? uri,
-          {dynamic data,
-          _i2.Options? options,
-          _i2.CancelToken? cancelToken,
-          _i2.ProgressCallback? onSendProgress,
-          _i2.ProgressCallback? onReceiveProgress}) =>
+  _i5.Future<_i2.Response<T>> putUri<T>(
+    Uri? uri, {
+    Object? data,
+    _i2.Options? options,
+    _i2.CancelToken? cancelToken,
+    _i2.ProgressCallback? onSendProgress,
+    _i2.ProgressCallback? onReceiveProgress,
+  }) =>
       (super.noSuchMethod(
-              Invocation.method(#putUri, [
-                uri
-              ], {
-                #data: data,
-                #options: options,
-                #cancelToken: cancelToken,
-                #onSendProgress: onSendProgress,
-                #onReceiveProgress: onReceiveProgress
-              }),
-              returnValue: Future<_i2.Response<T>>.value(_FakeResponse_6<T>()))
-          as _i5.Future<_i2.Response<T>>);
+        Invocation.method(
+          #putUri,
+          [uri],
+          {
+            #data: data,
+            #options: options,
+            #cancelToken: cancelToken,
+            #onSendProgress: onSendProgress,
+            #onReceiveProgress: onReceiveProgress,
+          },
+        ),
+        returnValue: _i5.Future<_i2.Response<T>>.value(_FakeResponse_6<T>(
+          this,
+          Invocation.method(
+            #putUri,
+            [uri],
+            {
+              #data: data,
+              #options: options,
+              #cancelToken: cancelToken,
+              #onSendProgress: onSendProgress,
+              #onReceiveProgress: onReceiveProgress,
+            },
+          ),
+        )),
+      ) as _i5.Future<_i2.Response<T>>);
   @override
-  _i5.Future<_i2.Response<T>> head<T>(String? path,
-          {dynamic data,
-          Map<String, dynamic>? queryParameters,
-          _i2.Options? options,
-          _i2.CancelToken? cancelToken}) =>
+  _i5.Future<_i2.Response<T>> head<T>(
+    String? path, {
+    Object? data,
+    Map<String, dynamic>? queryParameters,
+    _i2.Options? options,
+    _i2.CancelToken? cancelToken,
+  }) =>
       (super.noSuchMethod(
-              Invocation.method(#head, [
-                path
-              ], {
-                #data: data,
-                #queryParameters: queryParameters,
-                #options: options,
-                #cancelToken: cancelToken
-              }),
-              returnValue: Future<_i2.Response<T>>.value(_FakeResponse_6<T>()))
-          as _i5.Future<_i2.Response<T>>);
+        Invocation.method(
+          #head,
+          [path],
+          {
+            #data: data,
+            #queryParameters: queryParameters,
+            #options: options,
+            #cancelToken: cancelToken,
+          },
+        ),
+        returnValue: _i5.Future<_i2.Response<T>>.value(_FakeResponse_6<T>(
+          this,
+          Invocation.method(
+            #head,
+            [path],
+            {
+              #data: data,
+              #queryParameters: queryParameters,
+              #options: options,
+              #cancelToken: cancelToken,
+            },
+          ),
+        )),
+      ) as _i5.Future<_i2.Response<T>>);
   @override
-  _i5.Future<_i2.Response<T>> headUri<T>(Uri? uri,
-          {dynamic data, _i2.Options? options, _i2.CancelToken? cancelToken}) =>
+  _i5.Future<_i2.Response<T>> headUri<T>(
+    Uri? uri, {
+    Object? data,
+    _i2.Options? options,
+    _i2.CancelToken? cancelToken,
+  }) =>
       (super.noSuchMethod(
-              Invocation.method(#headUri, [uri],
-                  {#data: data, #options: options, #cancelToken: cancelToken}),
-              returnValue: Future<_i2.Response<T>>.value(_FakeResponse_6<T>()))
-          as _i5.Future<_i2.Response<T>>);
+        Invocation.method(
+          #headUri,
+          [uri],
+          {
+            #data: data,
+            #options: options,
+            #cancelToken: cancelToken,
+          },
+        ),
+        returnValue: _i5.Future<_i2.Response<T>>.value(_FakeResponse_6<T>(
+          this,
+          Invocation.method(
+            #headUri,
+            [uri],
+            {
+              #data: data,
+              #options: options,
+              #cancelToken: cancelToken,
+            },
+          ),
+        )),
+      ) as _i5.Future<_i2.Response<T>>);
   @override
-  _i5.Future<_i2.Response<T>> delete<T>(String? path,
-          {dynamic data,
-          Map<String, dynamic>? queryParameters,
-          _i2.Options? options,
-          _i2.CancelToken? cancelToken}) =>
+  _i5.Future<_i2.Response<T>> delete<T>(
+    String? path, {
+    Object? data,
+    Map<String, dynamic>? queryParameters,
+    _i2.Options? options,
+    _i2.CancelToken? cancelToken,
+  }) =>
       (super.noSuchMethod(
-              Invocation.method(#delete, [
-                path
-              ], {
-                #data: data,
-                #queryParameters: queryParameters,
-                #options: options,
-                #cancelToken: cancelToken
-              }),
-              returnValue: Future<_i2.Response<T>>.value(_FakeResponse_6<T>()))
-          as _i5.Future<_i2.Response<T>>);
+        Invocation.method(
+          #delete,
+          [path],
+          {
+            #data: data,
+            #queryParameters: queryParameters,
+            #options: options,
+            #cancelToken: cancelToken,
+          },
+        ),
+        returnValue: _i5.Future<_i2.Response<T>>.value(_FakeResponse_6<T>(
+          this,
+          Invocation.method(
+            #delete,
+            [path],
+            {
+              #data: data,
+              #queryParameters: queryParameters,
+              #options: options,
+              #cancelToken: cancelToken,
+            },
+          ),
+        )),
+      ) as _i5.Future<_i2.Response<T>>);
   @override
-  _i5.Future<_i2.Response<T>> deleteUri<T>(Uri? uri,
-          {dynamic data, _i2.Options? options, _i2.CancelToken? cancelToken}) =>
+  _i5.Future<_i2.Response<T>> deleteUri<T>(
+    Uri? uri, {
+    Object? data,
+    _i2.Options? options,
+    _i2.CancelToken? cancelToken,
+  }) =>
       (super.noSuchMethod(
-              Invocation.method(#deleteUri, [uri],
-                  {#data: data, #options: options, #cancelToken: cancelToken}),
-              returnValue: Future<_i2.Response<T>>.value(_FakeResponse_6<T>()))
-          as _i5.Future<_i2.Response<T>>);
+        Invocation.method(
+          #deleteUri,
+          [uri],
+          {
+            #data: data,
+            #options: options,
+            #cancelToken: cancelToken,
+          },
+        ),
+        returnValue: _i5.Future<_i2.Response<T>>.value(_FakeResponse_6<T>(
+          this,
+          Invocation.method(
+            #deleteUri,
+            [uri],
+            {
+              #data: data,
+              #options: options,
+              #cancelToken: cancelToken,
+            },
+          ),
+        )),
+      ) as _i5.Future<_i2.Response<T>>);
   @override
-  _i5.Future<_i2.Response<T>> patch<T>(String? path,
-          {dynamic data,
-          Map<String, dynamic>? queryParameters,
-          _i2.Options? options,
-          _i2.CancelToken? cancelToken,
-          _i2.ProgressCallback? onSendProgress,
-          _i2.ProgressCallback? onReceiveProgress}) =>
+  _i5.Future<_i2.Response<T>> patch<T>(
+    String? path, {
+    Object? data,
+    Map<String, dynamic>? queryParameters,
+    _i2.Options? options,
+    _i2.CancelToken? cancelToken,
+    _i2.ProgressCallback? onSendProgress,
+    _i2.ProgressCallback? onReceiveProgress,
+  }) =>
       (super.noSuchMethod(
-              Invocation.method(#patch, [
-                path
-              ], {
-                #data: data,
-                #queryParameters: queryParameters,
-                #options: options,
-                #cancelToken: cancelToken,
-                #onSendProgress: onSendProgress,
-                #onReceiveProgress: onReceiveProgress
-              }),
-              returnValue: Future<_i2.Response<T>>.value(_FakeResponse_6<T>()))
-          as _i5.Future<_i2.Response<T>>);
+        Invocation.method(
+          #patch,
+          [path],
+          {
+            #data: data,
+            #queryParameters: queryParameters,
+            #options: options,
+            #cancelToken: cancelToken,
+            #onSendProgress: onSendProgress,
+            #onReceiveProgress: onReceiveProgress,
+          },
+        ),
+        returnValue: _i5.Future<_i2.Response<T>>.value(_FakeResponse_6<T>(
+          this,
+          Invocation.method(
+            #patch,
+            [path],
+            {
+              #data: data,
+              #queryParameters: queryParameters,
+              #options: options,
+              #cancelToken: cancelToken,
+              #onSendProgress: onSendProgress,
+              #onReceiveProgress: onReceiveProgress,
+            },
+          ),
+        )),
+      ) as _i5.Future<_i2.Response<T>>);
   @override
-  _i5.Future<_i2.Response<T>> patchUri<T>(Uri? uri,
-          {dynamic data,
-          _i2.Options? options,
-          _i2.CancelToken? cancelToken,
-          _i2.ProgressCallback? onSendProgress,
-          _i2.ProgressCallback? onReceiveProgress}) =>
+  _i5.Future<_i2.Response<T>> patchUri<T>(
+    Uri? uri, {
+    Object? data,
+    _i2.Options? options,
+    _i2.CancelToken? cancelToken,
+    _i2.ProgressCallback? onSendProgress,
+    _i2.ProgressCallback? onReceiveProgress,
+  }) =>
       (super.noSuchMethod(
-              Invocation.method(#patchUri, [
-                uri
-              ], {
-                #data: data,
-                #options: options,
-                #cancelToken: cancelToken,
-                #onSendProgress: onSendProgress,
-                #onReceiveProgress: onReceiveProgress
-              }),
-              returnValue: Future<_i2.Response<T>>.value(_FakeResponse_6<T>()))
-          as _i5.Future<_i2.Response<T>>);
+        Invocation.method(
+          #patchUri,
+          [uri],
+          {
+            #data: data,
+            #options: options,
+            #cancelToken: cancelToken,
+            #onSendProgress: onSendProgress,
+            #onReceiveProgress: onReceiveProgress,
+          },
+        ),
+        returnValue: _i5.Future<_i2.Response<T>>.value(_FakeResponse_6<T>(
+          this,
+          Invocation.method(
+            #patchUri,
+            [uri],
+            {
+              #data: data,
+              #options: options,
+              #cancelToken: cancelToken,
+              #onSendProgress: onSendProgress,
+              #onReceiveProgress: onReceiveProgress,
+            },
+          ),
+        )),
+      ) as _i5.Future<_i2.Response<T>>);
   @override
-  void lock() => super.noSuchMethod(Invocation.method(#lock, []),
-      returnValueForMissingStub: null);
-  @override
-  void unlock() => super.noSuchMethod(Invocation.method(#unlock, []),
-      returnValueForMissingStub: null);
-  @override
-  void clear() => super.noSuchMethod(Invocation.method(#clear, []),
-      returnValueForMissingStub: null);
-  @override
-  _i5.Future<_i2.Response<dynamic>> download(String? urlPath, dynamic savePath,
-          {_i2.ProgressCallback? onReceiveProgress,
-          Map<String, dynamic>? queryParameters,
-          _i2.CancelToken? cancelToken,
-          bool? deleteOnError = true,
-          String? lengthHeader = r'content-length',
-          dynamic data,
-          _i2.Options? options}) =>
+  _i5.Future<_i2.Response<dynamic>> download(
+    String? urlPath,
+    dynamic savePath, {
+    _i2.ProgressCallback? onReceiveProgress,
+    Map<String, dynamic>? queryParameters,
+    _i2.CancelToken? cancelToken,
+    bool? deleteOnError = true,
+    String? lengthHeader = r'content-length',
+    Object? data,
+    _i2.Options? options,
+  }) =>
       (super.noSuchMethod(
-              Invocation.method(#download, [
-                urlPath,
-                savePath
-              ], {
-                #onReceiveProgress: onReceiveProgress,
-                #queryParameters: queryParameters,
-                #cancelToken: cancelToken,
-                #deleteOnError: deleteOnError,
-                #lengthHeader: lengthHeader,
-                #data: data,
-                #options: options
-              }),
-              returnValue: Future<_i2.Response<dynamic>>.value(
-                  _FakeResponse_6<dynamic>()))
-          as _i5.Future<_i2.Response<dynamic>>);
+        Invocation.method(
+          #download,
+          [
+            urlPath,
+            savePath,
+          ],
+          {
+            #onReceiveProgress: onReceiveProgress,
+            #queryParameters: queryParameters,
+            #cancelToken: cancelToken,
+            #deleteOnError: deleteOnError,
+            #lengthHeader: lengthHeader,
+            #data: data,
+            #options: options,
+          },
+        ),
+        returnValue:
+            _i5.Future<_i2.Response<dynamic>>.value(_FakeResponse_6<dynamic>(
+          this,
+          Invocation.method(
+            #download,
+            [
+              urlPath,
+              savePath,
+            ],
+            {
+              #onReceiveProgress: onReceiveProgress,
+              #queryParameters: queryParameters,
+              #cancelToken: cancelToken,
+              #deleteOnError: deleteOnError,
+              #lengthHeader: lengthHeader,
+              #data: data,
+              #options: options,
+            },
+          ),
+        )),
+      ) as _i5.Future<_i2.Response<dynamic>>);
   @override
-  _i5.Future<_i2.Response<dynamic>> downloadUri(Uri? uri, dynamic savePath,
-          {_i2.ProgressCallback? onReceiveProgress,
-          _i2.CancelToken? cancelToken,
-          bool? deleteOnError = true,
-          String? lengthHeader = r'content-length',
-          dynamic data,
-          _i2.Options? options}) =>
+  _i5.Future<_i2.Response<dynamic>> downloadUri(
+    Uri? uri,
+    dynamic savePath, {
+    _i2.ProgressCallback? onReceiveProgress,
+    _i2.CancelToken? cancelToken,
+    bool? deleteOnError = true,
+    String? lengthHeader = r'content-length',
+    Object? data,
+    _i2.Options? options,
+  }) =>
       (super.noSuchMethod(
-              Invocation.method(#downloadUri, [
-                uri,
-                savePath
-              ], {
-                #onReceiveProgress: onReceiveProgress,
-                #cancelToken: cancelToken,
-                #deleteOnError: deleteOnError,
-                #lengthHeader: lengthHeader,
-                #data: data,
-                #options: options
-              }),
-              returnValue: Future<_i2.Response<dynamic>>.value(
-                  _FakeResponse_6<dynamic>()))
-          as _i5.Future<_i2.Response<dynamic>>);
+        Invocation.method(
+          #downloadUri,
+          [
+            uri,
+            savePath,
+          ],
+          {
+            #onReceiveProgress: onReceiveProgress,
+            #cancelToken: cancelToken,
+            #deleteOnError: deleteOnError,
+            #lengthHeader: lengthHeader,
+            #data: data,
+            #options: options,
+          },
+        ),
+        returnValue:
+            _i5.Future<_i2.Response<dynamic>>.value(_FakeResponse_6<dynamic>(
+          this,
+          Invocation.method(
+            #downloadUri,
+            [
+              uri,
+              savePath,
+            ],
+            {
+              #onReceiveProgress: onReceiveProgress,
+              #cancelToken: cancelToken,
+              #deleteOnError: deleteOnError,
+              #lengthHeader: lengthHeader,
+              #data: data,
+              #options: options,
+            },
+          ),
+        )),
+      ) as _i5.Future<_i2.Response<dynamic>>);
   @override
-  _i5.Future<_i2.Response<T>> request<T>(String? path,
-          {dynamic data,
-          Map<String, dynamic>? queryParameters,
-          _i2.CancelToken? cancelToken,
-          _i2.Options? options,
-          _i2.ProgressCallback? onSendProgress,
-          _i2.ProgressCallback? onReceiveProgress}) =>
+  _i5.Future<_i2.Response<T>> request<T>(
+    String? path, {
+    Object? data,
+    Map<String, dynamic>? queryParameters,
+    _i2.CancelToken? cancelToken,
+    _i2.Options? options,
+    _i2.ProgressCallback? onSendProgress,
+    _i2.ProgressCallback? onReceiveProgress,
+  }) =>
       (super.noSuchMethod(
-              Invocation.method(#request, [
-                path
-              ], {
-                #data: data,
-                #queryParameters: queryParameters,
-                #cancelToken: cancelToken,
-                #options: options,
-                #onSendProgress: onSendProgress,
-                #onReceiveProgress: onReceiveProgress
-              }),
-              returnValue: Future<_i2.Response<T>>.value(_FakeResponse_6<T>()))
-          as _i5.Future<_i2.Response<T>>);
+        Invocation.method(
+          #request,
+          [path],
+          {
+            #data: data,
+            #queryParameters: queryParameters,
+            #cancelToken: cancelToken,
+            #options: options,
+            #onSendProgress: onSendProgress,
+            #onReceiveProgress: onReceiveProgress,
+          },
+        ),
+        returnValue: _i5.Future<_i2.Response<T>>.value(_FakeResponse_6<T>(
+          this,
+          Invocation.method(
+            #request,
+            [path],
+            {
+              #data: data,
+              #queryParameters: queryParameters,
+              #cancelToken: cancelToken,
+              #options: options,
+              #onSendProgress: onSendProgress,
+              #onReceiveProgress: onReceiveProgress,
+            },
+          ),
+        )),
+      ) as _i5.Future<_i2.Response<T>>);
   @override
-  _i5.Future<_i2.Response<T>> requestUri<T>(Uri? uri,
-          {dynamic data,
-          _i2.CancelToken? cancelToken,
-          _i2.Options? options,
-          _i2.ProgressCallback? onSendProgress,
-          _i2.ProgressCallback? onReceiveProgress}) =>
+  _i5.Future<_i2.Response<T>> requestUri<T>(
+    Uri? uri, {
+    Object? data,
+    _i2.CancelToken? cancelToken,
+    _i2.Options? options,
+    _i2.ProgressCallback? onSendProgress,
+    _i2.ProgressCallback? onReceiveProgress,
+  }) =>
       (super.noSuchMethod(
-              Invocation.method(#requestUri, [
-                uri
-              ], {
-                #data: data,
-                #cancelToken: cancelToken,
-                #options: options,
-                #onSendProgress: onSendProgress,
-                #onReceiveProgress: onReceiveProgress
-              }),
-              returnValue: Future<_i2.Response<T>>.value(_FakeResponse_6<T>()))
-          as _i5.Future<_i2.Response<T>>);
+        Invocation.method(
+          #requestUri,
+          [uri],
+          {
+            #data: data,
+            #cancelToken: cancelToken,
+            #options: options,
+            #onSendProgress: onSendProgress,
+            #onReceiveProgress: onReceiveProgress,
+          },
+        ),
+        returnValue: _i5.Future<_i2.Response<T>>.value(_FakeResponse_6<T>(
+          this,
+          Invocation.method(
+            #requestUri,
+            [uri],
+            {
+              #data: data,
+              #cancelToken: cancelToken,
+              #options: options,
+              #onSendProgress: onSendProgress,
+              #onReceiveProgress: onReceiveProgress,
+            },
+          ),
+        )),
+      ) as _i5.Future<_i2.Response<T>>);
   @override
   _i5.Future<_i2.Response<T>> fetch<T>(_i2.RequestOptions? requestOptions) =>
-      (super.noSuchMethod(Invocation.method(#fetch, [requestOptions]),
-              returnValue: Future<_i2.Response<T>>.value(_FakeResponse_6<T>()))
-          as _i5.Future<_i2.Response<T>>);
-  @override
-  String toString() => super.toString();
+      (super.noSuchMethod(
+        Invocation.method(
+          #fetch,
+          [requestOptions],
+        ),
+        returnValue: _i5.Future<_i2.Response<T>>.value(_FakeResponse_6<T>(
+          this,
+          Invocation.method(
+            #fetch,
+            [requestOptions],
+          ),
+        )),
+      ) as _i5.Future<_i2.Response<T>>);
 }
